+++ components/map/renderer/dimensions/point.ts
