+++ scripts/generate-map-geometry/example/index.ts
@@ -1,5 +1,5 @@
 import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
-import { fp } from "@/components/map/renderer/flexi-point";
+import { fp } from "@/components/map/renderer/dimensions/flexi-point";
 import { InterchangeBlueprint } from "@/scripts/generate-map-geometry/lib/interchange-blueprint";
 import { GeometryBuilder } from "@/scripts/generate-map-geometry/lib/geometry-builder";
 import { LineBuilder } from "@/scripts/generate-map-geometry/lib/line-builder";
