+++ scripts/generate-map-geometry/lib/line-builder.ts
@@ -1,6 +1,6 @@
 import { Segment } from "@/components/map/renderer/segment";
 import { LineColor } from "@/components/map/renderer/utils";
-import { FlexiPoint } from "@/components/map/renderer/flexi-point";
+import { FlexiPoint } from "@/components/map/renderer/dimensions/flexi-point";
 import { SegmentBuilder } from "@/scripts/generate-map-geometry/lib/segment-builder";
 import { SegmentInstruction } from "@/scripts/generate-map-geometry/lib/segment-instructions";
