+++ scripts/generate-map-geometry/lib/segment-builder.ts
@@ -2,7 +2,7 @@ import {
   flexi,
   FlexiLength,
 } from "@/components/map/renderer/dimensions/flexi-length";
-import { FlexiPoint } from "@/components/map/renderer/flexi-point";
+import { FlexiPoint } from "@/components/map/renderer/dimensions/flexi-point";
 import {
   CurveSegmentInstruction,
   SegmentInstruction,
