+++ scripts/generate-map-geometry/ptv/segments/flinders-street-to-richmond.ts
@@ -1,5 +1,5 @@
 import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
-import { FlexiPoint } from "@/components/map/renderer/flexi-point";
+import { FlexiPoint } from "@/components/map/renderer/dimensions/flexi-point";
 import {
   curve,
   SegmentInstruction,
