+++ scripts/generate-map-geometry/ptv/segments/southern-cross-to-north-melbourne.ts
@@ -1,5 +1,5 @@
 import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
-import { FlexiPoint } from "@/components/map/renderer/flexi-point";
+import { FlexiPoint } from "@/components/map/renderer/dimensions/flexi-point";
 import {
   diagonal,
   lineGap,
